+++ react-ecommerce/data.json
@@ -6171,6 +6171,12 @@
           "pinCode": "000077"
         }
       ]
+    },
+    {
+      "id": "5b17",
+      "email": "james@email.com",
+      "password": "Password@123",
+      "addresses": []
     }
   ],
   "cart": [],
